add a field to filter the recordings
The `RECORDINGS_IN_THIS_SESSION` field can be used to filter the recordings based on prior requirement.

It can be used for testing purposes (filtering out the test) or to create subtests (e.g. different subtests per room or source base angle).

Modificiations should be done so that the filtering is modulable